pyproject.toml
[tool.poetry]
name = "mptt"
version = "0.1.0"
description = ""
authors = ["Michael Gabbard <zanvoysan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
django-mptt = "^0.11.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
